migrations

+++ database/migrations/2022_05_23_155632_create_job_cards_table.php
@@ -15,7 +15,7 @@ class CreateJobCardsTable extends Migration
     {
         Schema::create('job_cards', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('customerid');
+            $table->integer('customerid');
             $table->integer('jobid');
             $table->string('title');
             $table->string('status');
